refactor: LadderGameResult 객체를 result 패키지에서 관리
