+++ src/main/java/ladder/domain/result/LadderGameResult.java
@@ -1,4 +1,4 @@
-package ladder.domain.ladder;
+package ladder.domain.result;
 
 import java.util.List;
