+++ src/main/java/ladder/view/OutputView.java
@@ -1,7 +1,7 @@
 package ladder.view;
 
 import ladder.domain.ladder.Ladder;
-import ladder.domain.ladder.LadderGameResult;
+import ladder.domain.result.LadderGameResult;
 import ladder.domain.ladder.LadderStep;
 import ladder.domain.ladder.Path;
 import ladder.domain.participant.Participant;
